requirements.txt
# AI Agent 核心依赖
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# 通义千问 SDK
dashscope>=1.20.0

# HTTP 客户端
httpx>=0.27.0

# 数据验证
pydantic>=2.0.0

# 环境配置
python-dotenv>=1.0.0

# 日志
loguru>=0.7.0

# RAG 向量检索
faiss-cpu>=1.8.0
langchain-text-splitters>=0.3.0

# DashScope Embeddings (通义千问向量模型)
dashscope>=1.20.0

# Web UI (Streamlit)
streamlit>=1.38.0
streamlit-ace>=0.1.1
streamlit-option-menu>=0.3.6
